refactor(web): extract the board InspectPanel into the file-viewer slice
Moves InspectPanel (dumb color/typography/spacing form, with its small
local draft-value state) and its InspectTarget/InspectStyleSnapshot/
InspectClickedDescendant UI types out of FileViewer.tsx and into the
file-viewer vertical slice, per ADR 0002. The orchestrator now imports
InspectPanel and its types from the slice barrel.
